# linux: optimize parseKeyValue