--- a/.changelog/186.txt
+++ b/.changelog/186.txt
@@ -1,0 +1,3 @@
+```release-note:enhancement
+linux: optimize linux key value parsing (ie: /proc files)
+```