As seen in this `alloc_objects` profile from auditbeat `linux.parseKeyValue()` accounts for a significant amount of the program allocations:
![profile003](https://github.com/elastic/go-sysinfo/assets/1712219/91ff72fd-936a-47b1-9d1c-f7035c24ee08)

This PR adds tests on the current implem to validate the expected behavior and proposes a new implementation for this function that removes _all_ allocs. As a nice side effect, it also cuts down cpu time by ~90%:

```
goos: linux
goarch: amd64
pkg: github.com/elastic/go-sysinfo/providers/linux
cpu: Intel(R) Core(TM) i7-10850H CPU @ 2.70GHz
                 │   old.bench   │              new.bench              │
                 │    sec/op     │   sec/op     vs base                │
ParseKeyValue-12   9169.0n ± 13%   984.8n ± 4%  -89.26% (p=0.000 n=10)

                 │  old.bench   │               new.bench               │
                 │     B/op     │     B/op      vs base                 │
ParseKeyValue-12   6.672Ki ± 0%   0.000Ki ± 0%  -100.00% (p=0.000 n=10)

                 │ old.bench  │             new.bench              │
                 │ allocs/op  │ allocs/op  vs base                 │
ParseKeyValue-12   58.00 ± 0%   0.00 ± 0%  -100.00% (p=0.000 n=10)
```