@andrewkroh 2023-07-26
A changelog entry will also be required. See https://github.com/elastic/go-sysinfo/blob/main/CONTRIBUTING.md.