@andrewkroh 2023-07-31
CI is failing on a missing license header. If you run `make update` it should add the header.