@andrewkroh 2023-08-02
Tests are still failing:

```
=== Failed
=== FAIL: providers/linux TestVmStatParse (0.00s)
    vmstat_test.go:165: unterminated input
```