@andrewkroh 2023-08-15
I compared the benchmark results from e4ac65ca824977 to f3a5024f2009. Looks good.

```
benchstat before.txt after.txt 
goos: darwin
goarch: arm64
pkg: github.com/elastic/go-sysinfo/providers/linux
                 │  before.txt   │              after.txt              │
                 │    sec/op     │    sec/op     vs base               │
ParseKeyValue-10   3540.0n ± ∞ ¹   972.6n ± ∞ ¹  -72.53% (p=0.008 n=5)
¹ need >= 6 samples for confidence interval at level 0.95

                 │  before.txt   │               after.txt               │
                 │     B/op      │     B/op       vs base                │
ParseKeyValue-10   6.672Ki ± ∞ ¹   0.000Ki ± ∞ ¹  -100.00% (p=0.008 n=5)
¹ need >= 6 samples for confidence interval at level 0.95

                 │ before.txt  │             after.txt              │
                 │  allocs/op  │ allocs/op   vs base                │
ParseKeyValue-10   58.00 ± ∞ ¹   0.00 ± ∞ ¹  -100.00% (p=0.008 n=5)
¹ need >= 6 samples for confidence interval at level 0.95
```